+++ resources/hpxml-measures/HPXMLtoOpenStudio/resources/weather.rb
@@ -357,7 +357,7 @@ class WeatherProcess
 
   def calc_ashrae_622_wsf(rowdata)
     require 'csv'
-    ashrae_csv = File.join(File.dirname(__FILE__), 'data_ashrae_622_wsf.csv')
+    ashrae_csv = File.join(File.dirname(__FILE__), 'data', 'ashrae_622_wsf.csv')
 
     wsf = nil
     CSV.read(ashrae_csv, headers: false).each do |data|
@@ -367,7 +367,7 @@ class WeatherProcess
     end
     return wsf unless wsf.nil?
 
-    # If not available in data_ashrae_622_wsf.csv...
+    # If not available in ashrae_622_wsf.csv...
     # Calculates the wSF value per report LBNL-5795E "Infiltration as Ventilation: Weather-Induced Dilution"
 
     # Constants